01 - Дополнительные правки по системе папок, вынос createElement и его подключение
