+++ index.html
@@ -1,14 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+<head>
+    <meta charset="UTF-8"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
     <link rel="stylesheet" href="./style.css">
     <title>Posts</title>
-  </head>
-  <body>
-    <h1>Posts</h1>
-    <div class="container-post"></div>
-    <script src="./script.js"></script>
-  </body>
+</head>
+<body>
+<h1>Posts</h1>
+<div class="container-post"></div>
+<script src="./utils/util.js"></script>
+<script type="module" src="./script/script.js"></script>
+</body>
 </html>
